# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tradosphere OS Gateway API
  version: "1.0.0"
  description: |-
    Unified API gateway for Tradosphere OS (Sprint 9, task 9.4). Implements
    the architecture decided in full in Decision D19 (EXECUTION_BOOK.md)
    before any gateway code was written.

    Two route families exist side by side:

    - Proxied (Auth, Market Data, Education, Portfolio, Analytics): the
      gateway reverse-proxies to each service's existing HTTP API via a
      hand-rolled wildcard handler that forwards the Authorization header
      verbatim. Each upstream service remains the sole auth authority for
      its own routes -- the gateway does NOT re-authenticate proxied
      requests.
    - In-process (Research, AI Council, CIO, Paper Trading, Journal): these
      five services are library-only with no HTTP surface of their own, so
      their routes are implemented directly on the gateway and authenticated
      by the gateway's own requireAuth (identical factory pattern to every
      other service in this repo).

    All domain routes are versioned under /v1/<service>/.... Infra routes
    (/health, /health/services, /metrics, /stream, /documentation,
    /openapi.yaml) are unversioned and, along with auth's own signup/login,
    are the only unauthenticated routes in the gateway.
servers:
  - url: http://localhost:4000
    description: Local gateway (apps/api)
...
security:
  - bearerAuth: []
paths:
...
  /v1/research/technical:
    post:
      tags: [Research]
      operationId: researchAnalyzeTechnical
      summary: Compute RSI/EMA/MACD, volume, and breakout signals from caller-supplied OHLCV bars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeTechnicalInput'
      responses:
        '200':
          description: "Either a computed TechnicalIndicatorSet (status: ok) or a ResearchGap (status: gap) if bars are insufficient."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TechnicalAnalysisResult'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/research/options:
    post:
      tags: [Research]
      operationId: researchAnalyzeOptions
      summary: Compute put/call ratio and open-interest-shift interpretation from a caller-supplied option chain snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionChainSnapshot'
      responses:
        '200':
          description: "Either a computed OptionChainAnalysis (status: ok) or a ResearchGap (status: gap) if the chain is empty or has no call-side open interest."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptionAnalysisResult'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/research/fundamentals/{symbol}:
    get:
      tags: [Research]
      operationId: researchFundamentals
      summary: Fetch the most recently ingested fundamentals verdict for a symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Either a computed FundamentalAnalysis (status: ok) or a ResearchGap (status: gap, reason: missing_fundamentals) if nothing has been ingested for this symbol."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundamentalAnalysisResult'
        '400':
          description: Malformed or empty symbol.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/research/sector:
    post:
      tags: [Research]
      operationId: researchAnalyzeSector
      summary: Compute relative strength and rotation for a sector against a benchmark from caller-supplied bars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeSectorInput'
      responses:
        '200':
          description: "Either a computed SectorAnalysis (status: ok) or a ResearchGap (status: gap, reason: missing_sector_data) if either series has fewer than 2 bars."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectorAnalysisResult'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/research/quant:
    post:
      tags: [Research]
      operationId: researchAnalyzeQuant
      summary: Compute a mean-reversion z-score and annualized volatility from caller-supplied bars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeQuantInput'
      responses:
        '200':
          description: "Either a computed QuantSignalSet (status: ok) or a ResearchGap (status: gap) if fewer than period+1 bars are supplied."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuantAnalysisResult'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/agents/technical:
    post:
      tags: [AI Council]
      operationId: aiAgentTechnical
      summary: Run the Technical expert agent over a technical analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalAnalysisResult'
      responses:
        '200':
          description: The Technical expert's opinion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertOpinion'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/agents/options:
    post:
      tags: [AI Council]
      operationId: aiAgentOptions
      summary: Run the Options expert agent over an option chain analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionAnalysisResult'
      responses:
        '200':
          description: The Options expert's opinion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertOpinion'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/agents/sector:
    post:
      tags: [AI Council]
      operationId: aiAgentSector
      summary: Run the Sector expert agent over a sector analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SectorAnalysisResult'
      responses:
        '200':
          description: The Sector expert's opinion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertOpinion'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/agents/quant:
    post:
      tags: [AI Council]
      operationId: aiAgentQuant
      summary: Run the Quant expert agent over a quant analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuantAnalysisResult'
      responses:
        '200':
          description: The Quant expert's opinion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertOpinion'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/ai/agents/fundamental:
    post:
      tags: [AI Council]
      operationId: aiAgentFundamental
      summary: Run the Fundamental expert agent over a fundamental analysis result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundamentalAnalysisResult'
      responses:
        '200':
          description: The Fundamental expert's opinion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertOpinion'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        Every proxied service verifies this itself against its own JWT
        secret; the gateway's own requireAuth (used only by the 10
        in-process routes) uses the identical verifyAccessToken /
        InvalidTokenError contract from @tradosphere/auth as every other
        service.
...